+++ src/main/java/BoardMapper.xml
@@ -4,11 +4,11 @@
   "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 <mapper namespace="BoardMapper">
 	<insert id = "insertBoard" parameterType="vo.BoardVO">
-		<selectKey resultType="_int" keyProperty="id" order="BEFORE">
-			select board_seq.nextval from dual
+		<selectKey resultType="_int" keyProperty="bid" order="BEFORE">
+			select BOARD_SEQ.nextval from dual
 		</selectKey>
 		insert into board(bid, writer, title, content, writedate, cnt)
-		values(#{bid},#{writer},#{title},#{content}, sysdate,#{cnt})
+		values(#{bid},#{writer},#{title},#{content}, sysdate,0)
 	</insert>
 	<update id="updateBoard" parameterType="vo.BoardVO">
 		update board set writer=#{writer},title=#{title},content=#{content},writedate= sysdate,cnt='0' where bid=#{bid}
