+++ src/main/java/my/spring/mini/BoardController.java
@@ -1,5 +1,7 @@
 package my.spring.mini;
 
+import javax.servlet.http.HttpSession;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -8,6 +10,7 @@ import org.springframework.web.servlet.ModelAndView;
 
 import dao.BoardDAO;
 import vo.BoardVO;
+import vo.Login_InfoVO;
 
 @Controller
 public class BoardController {
@@ -78,6 +81,32 @@ public class BoardController {
 			mav.addObject("listone", dao.listOne(Integer.valueOf(bid)));
 		}
 		mav.setViewName("boardContent");
+		return mav;
+	}
+	
+	@RequestMapping(value="/board/content/edit", method=RequestMethod.GET)
+	public ModelAndView doGetEdit(BoardVO vo, String action) {
+		ModelAndView mav = new ModelAndView();
+		
+		mav.setViewName("edit");
+		return mav;
+	}
+	@RequestMapping(value="/board/content/edit", method=RequestMethod.POST)
+	public ModelAndView doPostEdit(BoardVO vo, String action, String bid) {
+		ModelAndView mav = new ModelAndView();
+		if (action.equals("insert")) {
+			boolean result = dao.insert(vo);
+			if (result) {
+				mav.addObject("msg", "등록되었습니다.");
+			} else {
+				mav.addObject("mag", "등록에 실패하였습니다.");
+			}
+		}
+		else if(action.equals("update")) {
+			
+		}
+		mav.addObject("list", dao.listAll());
+		mav.setViewName("board");
 		return mav;
 	}
 }
