+++ src/main/java/my/spring/mini/RegisterController.java
@@ -1,0 +1,34 @@
+package my.spring.mini;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestMethod;
+import org.springframework.web.servlet.ModelAndView;
+
+import dao.UsersDAO;
+import vo.UsersVO;
+
+@Controller
+public class RegisterController {
+	@Autowired
+	UsersDAO dao;
+	
+	@RequestMapping(value="/register", method= RequestMethod.GET)
+	public String register() {
+		System.out.println("register");
+		return "register";
+	}
+	@RequestMapping(value="/register", method=RequestMethod.POST)
+	public ModelAndView doRegister(UsersVO vo) {
+		ModelAndView mav = new ModelAndView();
+		boolean result= dao.insert(vo);
+		if(result)
+			System.out.println("성공적으로 가입되었습니다.");
+		else
+			System.out.println(" 가입 실패하었습니다.");
+		
+		mav.setViewName("redirect:/");
+		return mav;
+	}
+}
